Implement mob spawning system/manager

+++ Lornworld2/Assets/Scripts/Shared/Entity.cs
@@ -5,8 +5,8 @@ public abstract class Entity : MonoBehaviour
     public ChunkPos ChunkPos => ChunkUtil.ToChunkPos(transform.position);
 
     [SerializeField]
-    protected int health;
+    protected float health;
 
     [SerializeField]
-    protected int maxHealth;
+    protected float maxHealth;
 }
